chore: run clippy

+++ library/container/mpegts/src/ts.rs
@@ -232,7 +232,7 @@ impl TsMuxer {
         ts_header.write_u8((stream_data.pid & 0xFF) as u8)?; //2
 
         /*continuity counter 4 bits*/
-        ts_header.write_u8(0x10 | (stream_data.continuity_counter & 0x0F) as u8)?; //3
+        ts_header.write_u8(0x10 | (stream_data.continuity_counter & 0x0F))?; //3
         stream_data.continuity_counter = (stream_data.continuity_counter + 1) % 16;
 
         if is_start {
